refactored connection methods
