+++ src/commands/workflow.rs
@@ -1,15 +1,9 @@
-use clap::{Args, Subcommand};
-use std::{error::Error, fs, io::Write, path::Path};
-
 use crate::{
-    cwl::{
-        clt::{CommandInputParameter, CommandLineTool},
-        format::format_cwl,
-        loader::{load_tool, load_workflow, resolve_filename},
-        wf::{Workflow, WorkflowOutputParameter},
-    },
+    cwl::{format::format_cwl, loader::load_workflow, wf::Workflow},
     io::{create_and_write_file, get_workflows_folder},
 };
+use clap::{Args, Subcommand};
+use std::{error::Error, fs, io::Write, path::Path};
 
 pub fn handle_workflow_commands(command: &WorkflowCommands) -> Result<(), Box<dyn Error>> {
     match command {
@@ -75,11 +69,11 @@ pub fn connect_workflow_nodes(args: &ConnectWorkflowArgs) -> Result<(), Box<dyn 
     let from_parts = args.from.split('/').collect::<Vec<_>>();
     let to_parts = args.to.split('/').collect::<Vec<_>>();
     if from_parts[0] == "@inputs" {
-        add_input_connection(from_parts[1], &args.to, &mut workflow)?;
+        workflow.add_input_connection(from_parts[1], &args.to)?;
     } else if to_parts[0] == "@outputs" {
-        add_output_connection(&args.from, to_parts[1], &mut workflow)?;
+        workflow.add_output_connection(&args.from, to_parts[1])?;
     } else {
-        step_connection(&args.from, &args.to, &mut workflow)?;
+        workflow.step_connection(&args.from, &args.to)?;
     }
 
     //save workflow
@@ -88,92 +82,6 @@ pub fn connect_workflow_nodes(args: &ConnectWorkflowArgs) -> Result<(), Box<dyn 
     let mut file = fs::File::create(&filename)?;
     file.write_all(yaml.as_bytes())?;
     println!("✔️  Updated Workflow {}!", filename);
-
-    Ok(())
-}
-
-/// Adds a connection between an input and a CommandLineTool. The tool will be registered as step if it is not already and an Workflow input will be added.
-pub fn add_input_connection(from_input: &str, to: &String, workflow: &mut Workflow) -> Result<(), Box<dyn Error>> {
-    let to_parts = to.split('/').collect::<Vec<_>>();
-
-    let to_filename = resolve_filename(to_parts[0]);
-    let to_tool: CommandLineTool = load_tool(&to_filename)?;
-    let to_slot = to_tool.inputs.iter().find(|i| i.id == to_parts[1]).expect("No slot");
-
-    //register input
-    if !workflow.has_input(from_input) {
-        workflow.inputs.push(CommandInputParameter::default().with_id(from_input).with_type(to_slot.type_.clone()));
-    }
-
-    workflow.add_new_step_if_not_exists(to_parts[0], &to_tool);
-    //add input in step
-    workflow
-        .steps
-        .iter_mut()
-        .find(|step| step.id == to_parts[0])
-        .unwrap()
-        .in_
-        .insert(to_parts[1].to_string(), from_input.to_owned());
-
-    println!("➕ Added or updated connection from inputs.{} to {} in workflow", from_input, to);
-
-    Ok(())
-}
-
-pub fn step_connection(from: &str, to: &str, workflow: &mut Workflow) -> Result<(), Box<dyn Error>> {
-    //handle from
-    let from_parts = from.split('/').collect::<Vec<_>>();
-    //check if step already exists and create if not
-    if !workflow.has_step(from_parts[0]) {
-        let from_filename = resolve_filename(from_parts[0]);
-        let from_tool: CommandLineTool = load_tool(&from_filename)?;
-        let from_outputs = from_tool.get_output_ids();
-        if !from_outputs.contains(&from_parts[1].to_string()) {
-            return Err(format!(
-                "❌ Tool {} does not have output `{}`. Cannot not create node from {} in Workflow!",
-                from_parts[0], from_parts[1], from_filename
-            )
-            .into());
-        }
-
-        //create step
-        workflow.add_new_step_if_not_exists(from_parts[0], &from_tool);
-    } else {
-        println!("🔗 Found step {} in workflow. Not changing that!", from_parts[0]);
-    }
-
-    //handle to
-    let to_parts = to.split('/').collect::<Vec<_>>();
-    //check if step exists
-    if !workflow.has_step(to_parts[0]) {
-        let to_filename = resolve_filename(to_parts[0]);
-        let to_tool: CommandLineTool = load_tool(&to_filename)?;
-
-        workflow.add_new_step_if_not_exists(to_parts[0], &to_tool);
-    }
-
-    let step = workflow.steps.iter_mut().find(|s| s.id == to_parts[0]).unwrap(); //safe here!
-    step.in_.insert(to_parts[1].to_string(), from.to_string());
-
-    Ok(())
-}
-
-pub fn add_output_connection(from: &String, to_output: &str, workflow: &mut Workflow) -> Result<(), Box<dyn Error>> {
-    let from_parts = from.split('/').collect::<Vec<_>>();
-
-    let from_filename = resolve_filename(from_parts[0]);
-    let from_tool: CommandLineTool = load_tool(&from_filename)?;
-    let from_slot = from_tool.outputs.iter().find(|i| i.id == from_parts[1]).expect("No slot");
-
-    if !workflow.has_output(to_output) {
-        workflow.outputs.push(WorkflowOutputParameter::default().with_id(to_output).clone());
-    }
-
-    let output = workflow.outputs.iter_mut().find(|o| o.id == to_output).unwrap();
-    output.type_ = from_slot.type_.clone();
-    output.output_source = from.clone();
-
-    println!("➕ Added or updated connection from {} to outputs.{} in workflow!", from, to_output);
 
     Ok(())
 }
