+++ src/cwl/wf.rs
@@ -1,8 +1,13 @@
-use std::collections::HashMap;
+use std::{collections::HashMap, error::Error};
 
 use serde::{Deserialize, Serialize};
 
-use super::{clt::{CommandInputParameter, CommandLineTool, Requirement}, types::CWLType};
+use crate::cwl::loader::{load_tool, resolve_filename};
+
+use super::{
+    clt::{CommandInputParameter, CommandLineTool, Requirement},
+    types::CWLType,
+};
 
 #[derive(Serialize, Deserialize, Debug, PartialEq)]
 #[serde(rename_all = "camelCase")]
@@ -62,6 +67,93 @@ impl Workflow {
             println!("➕ Added step {} to workflow", name);
         }
     }
+
+    /// Adds a connection between an input and a CommandLineTool. The tool will be registered as step if it is not already and an Workflow input will be added.
+    pub fn add_input_connection(self: &mut Self, from_input: &str, to: &String) -> Result<(), Box<dyn Error>> {
+        let to_parts = to.split('/').collect::<Vec<_>>();
+
+        let to_filename = resolve_filename(to_parts[0]);
+        let to_tool: CommandLineTool = load_tool(&to_filename)?;
+        let to_slot = to_tool.inputs.iter().find(|i| i.id == to_parts[1]).expect("No slot");
+
+        //register input
+        if !self.has_input(from_input) {
+            self.inputs.push(CommandInputParameter::default().with_id(from_input).with_type(to_slot.type_.clone()));
+        }
+
+        self.add_new_step_if_not_exists(to_parts[0], &to_tool);
+        //add input in step
+        self.steps
+            .iter_mut()
+            .find(|step| step.id == to_parts[0])
+            .unwrap()
+            .in_
+            .insert(to_parts[1].to_string(), from_input.to_owned());
+
+        println!("➕ Added or updated connection from inputs.{} to {} in workflow", from_input, to);
+
+        Ok(())
+    }
+
+    /// Adds a connection between an output and a CommandLineTool. The tool will be registered as step if it is not already and an Workflow output will be added.
+    pub fn add_output_connection(self: &mut Self, from: &String, to_output: &str) -> Result<(), Box<dyn Error>> {
+        let from_parts = from.split('/').collect::<Vec<_>>();
+
+        let from_filename = resolve_filename(from_parts[0]);
+        let from_tool: CommandLineTool = load_tool(&from_filename)?;
+        let from_slot = from_tool.outputs.iter().find(|i| i.id == from_parts[1]).expect("No slot");
+
+        if !self.has_output(to_output) {
+            self.outputs.push(WorkflowOutputParameter::default().with_id(to_output).clone());
+        }
+
+        let output = self.outputs.iter_mut().find(|o| o.id == to_output).unwrap();
+        output.type_ = from_slot.type_.clone();
+        output.output_source = from.clone();
+
+        println!("➕ Added or updated connection from {} to outputs.{} in workflow!", from, to_output);
+
+        Ok(())
+    }
+
+    /// Adds a connection between two a CommandLineToos. The tools will be registered as step if registered not already.
+    pub fn step_connection(self: &mut Self, from: &str, to: &str) -> Result<(), Box<dyn Error>> {
+        //handle from
+        let from_parts = from.split('/').collect::<Vec<_>>();
+        //check if step already exists and create if not
+        if !self.has_step(from_parts[0]) {
+            let from_filename = resolve_filename(from_parts[0]);
+            let from_tool: CommandLineTool = load_tool(&from_filename)?;
+            let from_outputs = from_tool.get_output_ids();
+            if !from_outputs.contains(&from_parts[1].to_string()) {
+                return Err(format!(
+                    "❌ Tool {} does not have output `{}`. Cannot not create node from {} in Workflow!",
+                    from_parts[0], from_parts[1], from_filename
+                )
+                .into());
+            }
+
+            //create step
+            self.add_new_step_if_not_exists(from_parts[0], &from_tool);
+        } else {
+            println!("🔗 Found step {} in workflow. Not changing that!", from_parts[0]);
+        }
+
+        //handle to
+        let to_parts = to.split('/').collect::<Vec<_>>();
+        //check if step exists
+        if !self.has_step(to_parts[0]) {
+            let to_filename = resolve_filename(to_parts[0]);
+            let to_tool: CommandLineTool = load_tool(&to_filename)?;
+
+            self.add_new_step_if_not_exists(to_parts[0], &to_tool);
+        }
+
+        let step = self.steps.iter_mut().find(|s| s.id == to_parts[0]).unwrap(); //safe here!
+        step.in_.insert(to_parts[1].to_string(), from.to_string());
+
+        Ok(())
+    }
 }
 
 #[derive(Serialize, Deserialize, Debug, Default, PartialEq, Clone)]
@@ -75,7 +167,7 @@ pub struct WorkflowStep {
 
 #[derive(Serialize, Deserialize, Debug, Default, PartialEq, Clone)]
 #[serde(rename_all = "camelCase")]
-pub struct WorkflowOutputParameter{
+pub struct WorkflowOutputParameter {
     pub id: String,
     pub type_: CWLType,
     pub output_source: String,
